Split up the dashboard utilities.
