+++ estimagic/tests/dashboard/test_monitoring_app.py
@@ -1,6 +1,8 @@
 """Test the functions of the monitoring app."""
 from pathlib import Path
 
+import numpy as np
+import pandas as pd
 from bokeh.document import Document
 from bokeh.models import ColumnDataSource
 
@@ -12,7 +14,10 @@ def test_monitoring_app():
     doc = Document()
     database_name = "test_db"
     current_dir_path = Path(__file__).resolve().parent
-    session_data = {"last_retrieved": 0, "database_path": current_dir_path / "db1.db"}
+    session_data = {
+        "last_retrieved": 0,
+        "database_path": current_dir_path / "test_databases" / "db1.db",
+    }
 
     monitoring.monitoring_app(
         doc=doc,
@@ -72,3 +77,87 @@ def test_set_last_retrieved_no_negative_value(monkeypatch):
     )
 
     assert session_data == {"last_retrieved": 0}
+
+
+# ====================================================================================
+# map_group_to_params
+# ====================================================================================
+
+
+def test_map_groups_to_params_group_none():
+    params = pd.DataFrame()
+    params["value"] = [0, 1, 2, 3]
+    params["group"] = None
+    params["name"] = ["a", "b", "c", "d"]
+    params.index = ["a", "b", "c", "d"]
+    expected = {}
+    res = monitoring._map_groups_to_params(params)
+    assert expected == res
+
+
+def test_map_groups_to_params_group_nan():
+    params = pd.DataFrame()
+    params["value"] = [0, 1, 2, 3]
+    params["group"] = np.nan
+    params["name"] = ["a", "b", "c", "d"]
+    params.index = ["a", "b", "c", "d"]
+    expected = {}
+    res = monitoring._map_groups_to_params(params)
+    assert expected == res
+
+
+def test_map_groups_to_params_group_empty():
+    params = pd.DataFrame()
+    params["value"] = [0, 1, 2, 3]
+    params["group"] = ["", "", "x", "x"]
+    params["name"] = ["a", "b", "c", "d"]
+    params.index = ["a", "b", "c", "d"]
+    expected = {"x": ["c", "d"]}
+    res = monitoring._map_groups_to_params(params)
+    assert expected == res
+
+
+def test_map_groups_to_params_group_false():
+    params = pd.DataFrame()
+    params["value"] = [0, 1, 2, 3]
+    params["group"] = [False, False, "x", "x"]
+    params["name"] = ["a", "b", "c", "d"]
+    params.index = ["a", "b", "c", "d"]
+    expected = {"x": ["c", "d"]}
+    res = monitoring._map_groups_to_params(params)
+    assert expected == res
+
+
+def test_map_groups_to_params_group_not_none():
+    params = pd.DataFrame()
+    params["value"] = [0, 1, 2, 3]
+    params["group"] = [None, "A", "B", "B"]
+    params.index = ["a", "b", "c", "d"]
+    params["name"] = ["a", "b", "c", "d"]
+    expected = {"A": ["b"], "B": ["c", "d"]}
+    res = monitoring._map_groups_to_params(params)
+    assert expected == res
+
+
+def test_map_groups_to_params_group_int_index():
+    params = pd.DataFrame()
+    params["value"] = [0, 1, 2, 3]
+    params.index = ["0", "1", "2", "3"]
+    params["name"] = ["0", "1", "2", "3"]
+    params["group"] = [None, "A", "B", "B"]
+    expected = {"A": ["1"], "B": ["2", "3"]}
+    res = monitoring._map_groups_to_params(params)
+    assert expected == res
+
+
+def test_map_groups_to_params_group_multi_index():
+    params = pd.DataFrame()
+    params["value"] = [0, 1, 2, 3]
+    params["group"] = [None, "A", "B", "B"]
+    params["ind1"] = ["beta", "beta", "cutoff", "cutoff"]
+    params["ind2"] = ["edu", "exp", 1, 2]
+    params.set_index(["ind1", "ind2"], inplace=True)
+    params["name"] = ["beta_edu", "beta_exp", "cutoff_1", "cutoff_2"]
+    expected = {"A": ["beta_exp"], "B": ["cutoff_1", "cutoff_2"]}
+    res = monitoring._map_groups_to_params(params)
+    assert expected == res
