+++ estimagic/tests/dashboard/test_run_dashboard.py
@@ -8,11 +8,12 @@ from estimagic.cli import cli
 from estimagic.dashboard.run_dashboard import _create_session_data
 from estimagic.dashboard.run_dashboard import _process_database_paths
 
+DATABASE_PATH = Path(__file__).resolve().parent / "test_databases"
+
 
 @pytest.fixture()
 def database_paths():
-    current_dir_path = Path(__file__).resolve().parent
-    database_paths = [current_dir_path / "db1.db", current_dir_path / "db2.db"]
+    database_paths = [DATABASE_PATH / "db1.db", DATABASE_PATH / "db2.db"]
     return database_paths
 
 
@@ -23,8 +24,7 @@ def database_name_to_path(database_paths):
 
 
 def test_process_dashboard_args_single_path():
-    current_dir_path = Path(__file__).resolve().parent
-    single_path = current_dir_path / "db1.db"
+    single_path = DATABASE_PATH / "db1.db"
     database_name_to_path = _process_database_paths(database_paths=single_path)
     assert database_name_to_path == {"db1": single_path}
 
@@ -73,7 +73,7 @@ def test_dashboard_cli(monkeypatch):
         cli,
         [
             "dashboard",
-            str(Path(__file__).parent / "*.db"),
+            str(DATABASE_PATH / "*.db"),
             "--no-browser",
             "--port",
             "9999",
@@ -99,9 +99,9 @@ def test_dashboard_cli_duplicate_paths(monkeypatch):
         cli,
         [
             "dashboard",
-            str(Path(__file__).parent / "*.db"),
-            str(Path(__file__).parent / "db1.db"),
-            str(Path(__file__).parent / "db2.db"),
+            str(DATABASE_PATH / "*.db"),
+            str(DATABASE_PATH / "db1.db"),
+            str(DATABASE_PATH / "db2.db"),
         ],
     )
 
@@ -117,6 +117,6 @@ def test_dashboard_cli_recursively_search_directories(monkeypatch):
     monkeypatch.setattr("estimagic.cli.run_dashboard", fake_run_dashboard)
 
     runner = CliRunner()
-    result = runner.invoke(cli, ["dashboard", str(Path(__file__).parent)])
+    result = runner.invoke(cli, ["dashboard", str(DATABASE_PATH)])
 
     assert result.exit_code == 0
